key_phrase_with_crawler/requirements_key_phrase_crawler.txt
azure-ai-textanalytics==5.1.0
azure-common==1.1.27
azure-core==1.20.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.7